Fix the bits argument of vssource.source()
The .source() function of the Indexer class will call initialize_clip
on the loaded clip, which converts 8-bit colors to 16-bit by default.

This means that the bits argument of vssource.source must be forwarded
to Indexer.source. Otherwise, loading an 8-bit source with bits=8 would
needlessly convert the colors to 16-bit and back.

Similarly, using bits=-1 to load a source in its native bit depth would
always return a 16-bit clip for an 8-bit source.