ExternalIndexer: Add filename suffix to idx file
The idea behind this is to make it easier to figure out which file belongs to what. This is especially useful for i.e. dgi files, in which case you may want to access the audio tracks of a specific video directly, or to pass the dgi file on to wibbly/wobbly for further processing in the case of DVDs.